ENH: Add --without|-W inverse of --within|-w
Yes, clients can negate the command's exit success, but this still doesn't handle whether the passed timespan is inclusive or exclusive. Because of the various time units that can be passed, subtracting one second would be complex at the client side, so let's better offer this. It also makes for a more fluent API without the negation.