Refactoring: Inline {write,increment}Marker() and implement --keep-context by omitting the CONTEXT column
Instead of the separate readMarker() invocation. This is more efficient and avoids concurrency issues in a non-transactional context. The fact that the CONTEXT column now won't be initialized is internal to the command and fully transparent to the user (an empty value will be returned, anyway).
By moving the --schema into database() (not needed for non-updates, but also doesn't do any harm), the three separate calls in update() are still DRY. Unfortunately, --update "$subject" cannot be passed to all three invocations (which would make for a updateMarker() extraction), because the (undocumented) feature to support multi-column contexts would then break.