ENH: Allow combination of --update with --get-{context,timestamp} and --with{in,out}
This is easy to implement, as the underlying miniDB now supports this and we just need to tack on the --columns * parameter and capture the output. Processing of the returned record is then delegated to the (now extracted) functionality from --query / --within / --without: parseMarker() followed by printAdditionalInformation() and calculateTimeRange().
This may help clients avoid a second query call, especially in combination with the new --increment-context feature.