Minor: Use ^G instead of ^A as the temporary placeholder
I have encountered problems with using ^A (but only ^A, not ^B) in the sparks implementation, when converting back to a newline. It does not affect this command, but to banish ^A as my go-to placeholder character, let's change it here as well. I hope nobody is crazy enough to use the bell inside a format string :-)